+++ src/main/scala/com/salsify/lineserver/shard/exception/KeyNotFoundException.scala
@@ -6,4 +6,4 @@ import com.salsify.lineserver.common.exception.LineServerException
   * Key not found exception.
   * @param key The key.
   */
-final case class KeyNotFoundException(key: Long) extends LineServerException(s"Key not found: '$key'")
+final case class KeyNotFoundException(key: Int) extends LineServerException(s"Key not found: '$key'")
